+++ src/main/java/xbc/service/MenuService.java
@@ -5,11 +5,11 @@ import java.util.Collection;
 import xbc.model.Menu;
 
 public interface MenuService {
-	public Menu findOne(long id);
+	public Menu findOne(Integer id);
 	public Collection<Menu> findAll();
 	public Menu update(Menu menu);
 	public void delete(Menu menu);
-	public void deleteById(long id);
+	public void deleteById(Integer id);
 	public void save(Menu menu);
 	public Collection<Menu> searchByTitle(String title);
 }
